mongodb -> aws S3 bucket

+++ src/main/java/com/mysite/sbb/answer/repository/AnswerRespository.java
@@ -3,5 +3,4 @@ package com.mysite.sbb.answer.repository;
 import com.mysite.sbb.answer.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnswerRespository extends JpaRepository<Answer, Long> {
-}
+public interface AnswerRespository extends JpaRepository<Answer, Integer> { }
